Initial Commit
Much to be fixed. In an attempt to refactor the js code I broke the animation.
Now it is not adding the class to the elements as desired. Next commit will
need to fix this. There is no content yet for this webpage. Most of this
commit dealt with the main animation.