Fixed Animation and Added to All Elements
The animation which I had broken is now back to working. Also, I added the main
animation to all of the elements so that it goes to the designate section when
clicking on each of the divs. I need to add the selected animation to About and
Portfolio, the animations moving between sections, and of course content for each
of the sections.